refactor: move files related to RawFluxDataTable
