+++ giraffe/src/components/RawFluxDataTable/index.ts
@@ -1,0 +1,1 @@
+export {RawFluxDataTable} from './RawFluxDataTable'
